AU-0 - Updating transfer-artifacts-in-sub_folders_in_parallel

+++ transfer-artifacts-in-sub_folders_in_parallel/sv_test_migrate_folder.sh
@@ -79,12 +79,52 @@ run_migrate_command() {
         elif [ "${TRANSFERONLY}" = "yes" ]; then
             while IFS= read -r line
             do
-                echo "jf rt dl \"$source_repo/$line\" . --threads=8 --server-id $source_artifactory ; jf rt u \"$line\" \"$target_repo/$line\" --threads=8 --server-id $target_artifactory ; rm -rf \"$line\" "$'\n'
-                #jf rt dl \"$1/$line\" . --server-id $SOURCE_ID ; jf rt u \"$line\" \"$1/$line\" --server-id $TARGET_ID ; rm -f \"$line\"
+
+                # Does the artifact have properties
+                get_item_properties_cmd="jf rt  curl -s -k -XGET \"/api/storage/$source_repo/$line?properties\" --server-id $source_artifactory"
+                # echo $get_item_properties_cmd
+                prop_output=$(eval "$get_item_properties_cmd")
+                prop_exit_status=$?
+                # echo $prop_output
+
+                if [ $prop_exit_status -ne 0 ]; then
+                    echo "Error: Command failed for folder: $folder_to_migrate - Run Command: $get_item_properties_cmd"
+                    echo "Error: Command failed for folder: $folder_to_migrate - Run Command: $get_item_properties_cmd" >> "$failed_commands_file"
+                else
+                    # Check the status code and process the JSON data
+                    http_status=$(echo "$prop_output" | jq -r '.errors[0].status')
+                    if [ "$http_status" != "404" ]; then
+                       # The artifact has properties
+                        json_data=$(echo "$prop_output" | jq -c '.properties')
+                        #escaped_json_data=$(echo "$json_data" | sed 's/"/\\"/g')
+                        # Construct the modified JSON data dynamically
+                        #modified_json="{\\"props\\": $escaped_json_data}"
+                         modified_json="{\"props\": $json_data}"
+                         escaped_modified_json=$(echo "$modified_json" | sed 's/"/\\"/g')
+                        # Run the PATCH request using the modified JSON data to set the properties for the artifact after upload
+                        #echo "jf rt dl \"$source_repo/$line\" . --threads=8 --server-id $source_artifactory ; jf rt u \"$line\" \"$target_repo/$line\" --threads=8 --server-id $target_artifactory ; rm -rf \"$line\" "$'\n'
+                        # Cannot apply the XPATCH  to  /api/metadata as indexing takes time. See https://jfrog-int.atlassian.net/browse/RTDEV-34900 . 
+
+                        # echo "jf rt curl -k -sL -XPATCH -H 'Content-Type: application/json' '/api/metadata/$target_repo/$line?atomicProperties=1' --server-id $target_artifactory -d '$modified_json'"$'\n' >> properties_patch.txt
+
+                        echo "jf rt curl -k -sL -XPATCH -H \"Content-Type: application/json\" \"/api/metadata/$target_repo/$line?atomicProperties=1\" --server-id $target_artifactory -d \"$escaped_modified_json\""$'\n' >> properties_patch.txt
+                        # So run the commands in properties_patch.txt in the end after all the artifacts are uploaded to target.
+
+                        echo "jf rt dl \"$source_repo/$line\" . --threads=8 --server-id $source_artifactory ; jf rt u \"$line\" \"$target_repo/$line\" --threads=8 --server-id $target_artifactory ; jf rt curl -k -sL -XPATCH -H \"Content-Type: application/json\" \"/api/metadata/$target_repo/$line?atomicProperties=1\" --server-id $target_artifactory -d \"$escaped_modified_json\"; rm -rf \"$line\" "$'\n' 
+                        #eval "jf rt curl -k -sL -XPATCH -H 'Content-Type: application/json' '/api/metadata/$target_repo/$line?atomicProperties=1' --server-id $target_artifactory -d '$modified_json'"
+
+                    else
+                       # The artifact does not have properties
+                        echo "jf rt dl \"$source_repo/$line\" . --threads=8 --server-id $source_artifactory ; jf rt u \"$line\" \"$target_repo/$line\" --threads=8 --server-id $target_artifactory ; rm -rf \"$line\" "$'\n'
+                    fi
+
+                fi
+
             done < "c"
         else 
             echo "Wrong 5th Parameter, 5th parameter value should be yes or no"
         fi 
+        
         rm -f a b c
     fi
 }
@@ -319,4 +359,6 @@ if [ -d "$root_folder" ]; then
         # Add additional actions for non-empty folders here if needed
     fi
 fi
+
+# echo "Next PATCH the properties using patch_properties.sh "
